use PostConstruct
